chore: restructure layout of repository
This began as a pedantic restructuring to sort-of dry out the code but
turned into a silly, yet major overhaul. No functionality was fundamentally
changed, added, or removed. Only minor changes were made otherwise
(using prepared statements in sql queries, for example).

One major advantage to this pattern is the condensing of the enumeration
and orchestration modules into a single module (read a single
`go mod tidy` on clone will get one up and running). This change also
explicitly defines the module/packages.

Again, this was all just a silly exercise. I absolutely will not be
offended if this is closed with prejudice (obliterated and never
addressed again). I acknowledge this PR likely serves no purpose towards
the goal of this repo and that my time could have been spent elsewhere.

Squashed commit of the following:

commit ac475a768fe4486fb47ecfdf1b2743adf78f625f
Author: greg linton <2653109+glinton@users.noreply.github.com>
Date:   Thu Jan 12 16:01:27 2023 -0700

    chore: port existing tests to correct package, add test docker-compose file

commit b27878e8bc0c49164b0403da0c60827fef52f97f
Author: greg linton <2653109+glinton@users.noreply.github.com>
Date:   Wed Jan 11 17:43:26 2023 -0700

    chore: update gitignore

commit a30a1191e719e7b3a7b235aa0097403602b8c504
Author: greg linton <2653109+glinton@users.noreply.github.com>
Date:   Wed Jan 11 17:42:00 2023 -0700

    chore: clean up some remaining error bits

    enumerate exits rather than returns errors

commit 469e2b0d52c05daede3b25a2aa546bc9057931bc
Author: greg linton <2653109+glinton@users.noreply.github.com>
Date:   Wed Jan 11 17:39:35 2023 -0700

    chore: minor cleanups, sql safety, and commenting

commit b66db775c725985477f7532fb8643a2b0b2a0f9d
Author: greg linton <2653109+glinton@users.noreply.github.com>
Date:   Wed Jan 11 16:23:27 2023 -0700

    chore: more tidying orchestrator

commit c7a543fda14b486b8c17748830e9e4c432176621
Author: greg linton <2653109+glinton@users.noreply.github.com>
Date:   Wed Jan 11 13:48:17 2023 -0700

    chore: sql tidy

commit 19d26abc3cce74593232df00ed7ad88136b2bc9e
Author: greg linton <2653109+glinton@users.noreply.github.com>
Date:   Mon Jan 9 15:50:30 2023 -0700

    chore: more cleanup

commit 80d926ec79bfd949a27f720db2d6e1ca4d1ea3f8
Author: greg linton <2653109+glinton@users.noreply.github.com>
Date:   Mon Jan 9 12:39:42 2023 -0700

    chore: tidy and restructure codebase